+++ web/src/components/molecules/Common/Header/index.tsx
@@ -1,7 +1,6 @@
 import styled from "@emotion/styled";
 import { MenuProps } from "antd";
-import { useCallback, useMemo } from "react";
-import { useNavigate } from "react-router-dom";
+import { useMemo } from "react";
 
 import { useAuth } from "@reearth-cms/auth";
 import Header from "@reearth-cms/components/atoms/Header";
@@ -23,6 +22,8 @@ export interface Props {
   currentProject?: Project;
   onWorkspaceModalOpen: () => void;
   onNavigateToSettings: () => void;
+  onWorkspaceNavigation: (id: number) => void;
+  onHomeNavigation: () => void;
   logoUrl?: string;
 }
 
@@ -34,11 +35,12 @@ const HeaderMolecule: React.FC<Props> = ({
   currentProject,
   onWorkspaceModalOpen,
   onNavigateToSettings,
+  onWorkspaceNavigation,
+  onHomeNavigation,
   logoUrl,
 }) => {
   const t = useT();
   const { logout } = useAuth();
-  const navigate = useNavigate();
   const url = useMemo(() => {
     if (window.REEARTH_CONFIG?.editorUrl && currentWorkspace?.id) {
       return new URL(`dashboard/${currentWorkspace.id}`, window.REEARTH_CONFIG?.editorUrl);
@@ -50,17 +52,6 @@ const HeaderMolecule: React.FC<Props> = ({
     () => currentWorkspace?.id === personalWorkspace?.id,
     [currentWorkspace?.id, personalWorkspace?.id],
   );
-
-  const handleWorkspaceNavigation = useCallback(
-    (id: number) => {
-      navigate(`/workspace/${id}`);
-    },
-    [navigate],
-  );
-
-  const handleHomeNavigation = useCallback(() => {
-    navigate(`/workspace/${currentWorkspace?.id}`);
-  }, [currentWorkspace?.id, navigate]);
 
   const WorkspacesItems: MenuProps["items"] = useMemo(
     () => [
@@ -79,7 +70,7 @@ const HeaderMolecule: React.FC<Props> = ({
             key: workspace.id,
             icon: <UserAvatar username={workspace.name} size="small" />,
             style: { paddingLeft: 0, paddingRight: 0 },
-            onClick: () => handleWorkspaceNavigation(workspace.id),
+            onClick: () => onWorkspaceNavigation(workspace.id),
           })),
       },
       {
@@ -100,7 +91,7 @@ const HeaderMolecule: React.FC<Props> = ({
             key: workspace.id,
             icon: <UserAvatar username={workspace.name} size="small" shape="square" />,
             style: { paddingLeft: 0, paddingRight: 0 },
-            onClick: () => handleWorkspaceNavigation(workspace.id),
+            onClick: () => onWorkspaceNavigation(workspace.id),
           })),
       },
       {
@@ -110,7 +101,7 @@ const HeaderMolecule: React.FC<Props> = ({
         onClick: onWorkspaceModalOpen,
       },
     ],
-    [t, onWorkspaceModalOpen, handleWorkspaceNavigation, workspaces, personalWorkspace],
+    [t, workspaces, onWorkspaceModalOpen, personalWorkspace?.id, onWorkspaceNavigation],
   );
 
   const AccountItems: MenuProps["items"] = useMemo(
@@ -134,9 +125,9 @@ const HeaderMolecule: React.FC<Props> = ({
   return (
     <MainHeader>
       {logoUrl ? (
-        <LogoIcon src={logoUrl} onClick={handleHomeNavigation} />
+        <LogoIcon src={logoUrl} onClick={onHomeNavigation} />
       ) : (
-        <Logo onClick={handleHomeNavigation}>{t("Re:Earth CMS")}</Logo>
+        <Logo onClick={onHomeNavigation}>{t("Re:Earth CMS")}</Logo>
       )}
       <VerticalDivider />
       <WorkspaceDropdown
